Parameter Make.1-P

type t

Type of the nodes as specified by the user

type id

Node identifier that can be compared and hashed using the generic comparison and hashing functions of the standard library. Typically an int or a string.

val id : t -> id

User function to extract a node's unique identifier

val to_string : id -> string

User function to make a node identifier printable, used for debugging only.